pyproject.toml
[project]
name = "fastapi-user-service"
version = "0.1.0"
description = ""
authors = [
    {name = "zhangyu528",email = "zhangyu528@163.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "pydantic[email] (>=1.10.13,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "fastapi-jwt-auth (>=0.5.0,<0.6.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
uvicorn = "^0.34.2"
alembic = "^1.15.2"
pytest = "^8.3.5"
httpx = "^0.28.1"

